@@ -1,15 +1,4 @@
-class DocumentsController < ApplicationController
-  # GET /documents
-  # GET /documents.json
-  def index
-    @documents = Document.all
-
-    respond_to do |format|
-      format.html # index.html.erb
-      format.json { render json: @documents }
-    end
-  end
-
+class DocumentsController < InheritedResources::Base
   def search
     @documents = Document.basic_search(params[:query])
 
@@ -19,40 +8,9 @@ class DocumentsController < ApplicationController
     end
   end
 
-  # GET /documents/1
-  # GET /documents/1.json
-  def show
-    @document = Document.find(params[:id])
-
-    respond_to do |format|
-      format.html # show.html.erb
-      format.json { render json: @document }
-    end
-  end
-
-  # GET /documents/new
-  # GET /documents/new.json
-  def new
-    @document = Document.new
-
-    respond_to do |format|
-      format.html # new.html.erb
-      format.json { render json: @document }
-    end
-  end
-
-  # GET /documents/1/edit
-  def edit
-    @document = Document.find(params[:id])
-  end
-
-  # POST /documents
-  # POST /documents.json
   def create
-    @document = Document.new(params[:document])
-
-    respond_to do |format|
-      if @document.save && @document.extract_content
+    create! do |format|
+      if @document.errors.empty? && @document.extract_content
         format.html { redirect_to documents_path, notice: 'Document was successfully created.' }
         format.json { render json: @document, status: :created, location: @document }
       else
@@ -62,13 +20,9 @@ class DocumentsController < ApplicationController
     end
   end
 
-  # PUT /documents/1
-  # PUT /documents/1.json
   def update
-    @document = Document.find(params[:id])
-
-    respond_to do |format|
-      if @document.update_attributes(params[:document]) && @document.extract_content
+    update! do |format|
+      if @document.errors.empty? && @document.extract_content
         format.html { redirect_to document_path notice: 'Document was successfully updated.' }
         format.json { head :no_content }
       else
@@ -78,15 +32,9 @@ class DocumentsController < ApplicationController
     end
   end
 
-  # DELETE /documents/1
-  # DELETE /documents/1.json
-  def destroy
-    @document = Document.find(params[:id])
-    @document.destroy
+  private 
 
-    respond_to do |format|
-      format.html { redirect_to documents_url }
-      format.json { head :no_content }
-    end
+  def permitted_params
+    params.permit(:document => [:title, :doc])
   end
 end
